Issue #14: The input type of the launch configuration does not exist - eclipse
